adding ability to display references in logic view. This pulls the references out of the field values array where they are in our internal representation and places them in their own array for processing and displaying in the logic views
Adding unique criteria id to each source data criteira instance. This is required to keep track of references between the data criteria as the patient record is essencially wiped out and rebuilt based on updated source data criteria objects.

Adding round trip saving of references and materializing the reference in the HDS patient model

Updated templates to display referenced critiera description and start and end dates.

TODO -- and automatic updating of references in the patient builder when a referenced criteria is either removed or the dates are updated.